+++ modules/payu/payment_succcess.php
@@ -1,10 +1,10 @@
 <?php
 /**
-*	ver. 0.1.5.1
+*	ver. 1.6
 *	PayU Payment Modules
 *	
 *	@copyright  Copyright 2012 by PayU
-*	@license    http://opensource.org/licenses/LGPL-3.0  Open Software License (LGPL 3.0)
+*	@license    http://opensource.org/licenses/GPL-3.0  Open Software License (GPL 3.0)
 *	http://www.payu.com
 *	http://twitter.com/openpayu
 */
@@ -12,6 +12,15 @@ $useSSL = true;
 include(dirname(__FILE__).'/../../config/config.inc.php');
 include(dirname(__FILE__).'/../../header.php');
 ob_clean();
+
 $payu = new Payu();
+$values = array('order' => Tools::getValue('order'), 'error' => Tools::getValue('error'));
+
+if(Tools::getValue('error'))
+{
+    Tools::redirectLink($payu->getModuleAddress(true, true) . 'payment_error.php?'.http_build_query($values, '', '&'));
+    exit;
+}
+
 $payu->execSuccessOrder((int)(Tools::getValue('order')));
 exit();